requirements.txt
aiogram>=3.0.0
fastapi>=0.110
uvicorn>=0.27
pytest-asyncio>=0.23
